refactor device flow into oauth flow
